requirements.txt
pandas>=2.0.0
osmnx>=1.3.0
networkx>=3.0
geopy>=2.3.0
pyyaml>=6.0.1
tqdm>=4.65.0
scikit-learn>=1.3.0
folium>=0.14.0
Pillow>=10.0.0
piexif>=1.1.3
